# phase-5/specs/2-advanced-cloud-deployment/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Chatbot Chat API
  version: 1.0.0
  description: |
    REST API for the Todo Chatbot Chat API service providing task CRUD operations,
    tag management, and health checks. All task operations publish events to Kafka
    via Dapr Pub/Sub.
  contact:
    name: Todo Chatbot Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: http://chat-api.todo.svc.cluster.local:8000
    description: Kubernetes cluster
security:
  - bearerAuth: []
paths:
...
  /tags:
    get:
      summary: List all tags
      operationId: listTags
      tags:
        - Tags
      parameters:
        - name: search
          in: query
          description: Filter tags by name prefix
          schema:
            type: string
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      $ref: '#/components/schemas/TagResponse'

    post:
      summary: Create a new tag
      operationId: createTag
      tags:
        - Tags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  pattern: '^[a-z0-9-]+$'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '409':
          description: Tag already exists
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TagResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        usage_count:
          type: integer
...